% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-load.r
\name{in_locate_symbol}
\alias{in_locate_symbol}
\title{Locate which dataset(s) contain the specified symbols}
\usage{
in_locate_symbol(Symbol = NULL)
}
\arguments{
\item{Symbol}{Character vector of symbols to locate, or NULL}
}
\value{
Character vector containing "Q", "Y", both, or NULL
}
\description{
Locate which dataset(s) contain the specified symbols
}
\keyword{internal}
